+++ lib/sequel/adapters/jdbc/h2.rb
@@ -130,7 +130,7 @@ module Sequel
             sql = 'SELECT IDENTITY();'
             rs = log_yield(sql){stmt.executeQuery(sql)}
             rs.next
-            rs.getInt(1)
+            rs.getLong(1)
           end
         end
