+++ lib/sequel/adapters/jdbc/oracle.rb
@@ -72,7 +72,7 @@ module Sequel
               begin
                 rs = log_yield(sql){stmt.executeQuery(sql)}
                 rs.next
-                rs.getInt(1)
+                rs.getLong(1)
               rescue java.sql.SQLException
                 nil
               end
